fix(recommendations): stop rounding the recommended latency to whole milliseconds
The column was INTEGER and the controller rounded to match, while the download and upload beside it have always been stored to two decimals - so the one figure of the three measured in fractions was the one stored without them.

On a fibre or a local line that is most of the reading: 0.4 ms and 1.4 ms produced the same recommendation. Anything under half a millisecond became 0, and 0 is the damaging value rather than merely an imprecise one - it reads as the best latency ever measured on the line, and createRecommendations only replaces the stored figure with something lower, so no later test can beat it and the recommendation stays wrong for the life of the database.

Migration 0012 follows 0010, which made the same change to speedtests.ping, including why sqlite is skipped: changeColumn there rebuilds the table from describeTable's output, which does not report autoIncrement, and sqlite's numeric affinity stores a fractional value as REAL regardless.

+++ server/models/Recommendations.js
@@ -3,7 +3,12 @@ import db from '../config/database.js';
 
 export default db.define("recommendations", {
     ping: {
-        type: Sequelize.INTEGER,
+        // DOUBLE, as the two beside it and as speedtests.ping - see migration
+        // 0012. An INTEGER here rounded away the fraction that is most of a
+        // latency reading on a fast line, and a sub-millisecond one became 0 -
+        // which reads as the best latency ever measured and so can never be
+        // improved on by a later test.
+        type: Sequelize.DOUBLE,
         allowNull: false
     },
     download: {
